:bug: Viewer ReferenceError: navigator is not defined 오류 수정
